pyproject.toml
[project]
name = "textual-jsx"
version = "0.1.0"
description = "Example library to enable JSX support in Textual"
authors = [
    {name = "Frost Ming", email = "me@frostming.com"},
]
dependencies = [
    "textual-dev>=1.8.0",
    "python-jsx>=0.4.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
